refactor(animation): move animations to its own mod
First step for more rusty version animations. The goal is to make
animatnion more generic so its easier to add new animations to komorebi!
